refactor - RacingCarList를 가지는 일급 콜렉션으로 추출하고 리펙토링
