+++ src/main/java/step3/RacingCarManager.java
@@ -2,43 +2,26 @@ package step3;
 
 import step3.strategy.MoveStrategy;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class RacingCarManager {
-    private List<Car> joinCarList;
-    private MoveStrategy moveStrategy;
 
-    public RacingCarManager(MoveStrategy moveStrategy) {
+    private MoveStrategy moveStrategy;
+    private RacingCarList racingCarList;
+
+    public RacingCarManager(int carCount, MoveStrategy moveStrategy) {
         if (moveStrategy == null) {
             throw new NullPointerException(ErrorMessage.MOVE_STRATEGY_IS_NULL);
         }
         this.moveStrategy = moveStrategy;
-    }
-
-    public void joinCars(int carCount) {
-        List<Car> list = new ArrayList<>();
-        IntStream.range(0, carCount).forEach(i -> {
-            list.add(new Car());
-        });
-        joinCarList = Collections.unmodifiableList(list);
+        racingCarList = new RacingCarList(carCount);
     }
 
     public void moveCars() {
-        joinCarList
-                .forEach(car -> {
-                    int generate = moveStrategy.generate();
-                    car.move(generate);
-                });
+        racingCarList.moveCars(moveStrategy);
     }
 
-    public List<Integer> getRacingStatus() {
-        return joinCarList
-                .stream()
-                .map(Car::getLocation)
-                .collect(Collectors.toList());
+    public RacingCarList getRacingCarList() {
+        return racingCarList;
     }
+
 }
